casting strlen to unsigned int

+++ singly_linked_lists/2-add_node.c
@@ -15,7 +15,7 @@ list_t *add_node(list_t **head, const char *str)
 
 	new->str = strdup(str);
 	new->next = *head;
-	new->len = strlen(str);
+	new->len = (unsigned int)strlen(str);
 	*head = new;
 	return (new);
 }
